feat: Enhanced orderbook system with Net Price formula
- Removed OTC row and related functionality
- Added Raw Price and Net Price columns (8 columns total)
- Implemented Net Price formula: (Raw Price × Input) / (Input + KDV)
- Updated commission calculations to use float values
- Added Binance disabled display with opacity styling
- Updated HTML template with Raw Price and Net Price headers
- Cleaned up cursor position management and OTC calculation code
- Enhanced tooltips for better price calculation transparency

+++ app/core/config.py
@@ -28,9 +28,9 @@ MIN_CHANGE_AMOUNT = 0.001  # Minimum change amount to record
 DUST_FILTER_THRESHOLD = 0.000001  # Filter out dust transactions
 
 # Exchange Configuration
-BINANCE_COMMISSION_BPS = int(os.getenv("BINANCE_COMMISSION_BPS", "10"))  # 10 bps = 0.1%
-COINTR_COMMISSION_BPS = int(os.getenv("COINTR_COMMISSION_BPS", "15"))    # 15 bps = 0.15%
-WHITEBIT_COMMISSION_BPS = int(os.getenv("WHITEBIT_COMMISSION_BPS", "10")) # 10 bps = 0.1%
+BINANCE_COMMISSION_BPS = float(os.getenv("BINANCE_COMMISSION_BPS", "10"))  # 10 bps = 0.1%
+COINTR_COMMISSION_BPS = float(os.getenv("COINTR_COMMISSION_BPS", "15"))    # 15 bps = 0.15%
+WHITEBIT_COMMISSION_BPS = float(os.getenv("WHITEBIT_COMMISSION_BPS", "10")) # 10 bps = 0.1%
 KDV_RATE = float(os.getenv("KDV_RATE", "0.20"))                         # 20% KDV
 
 # Binance API Configuration
